+++ .github/workflows/test_1.x_java.yml
@@ -32,4 +32,4 @@ jobs:
         distribution: 'zulu'
         java-package: jdk
     - name: test with Maven
-      run: mvn clean test -Pv1x-tests
+      run: mvn clean test -Pschema-v1-tests
