+++ .github/workflows/test_1.x_php.yml
@@ -14,7 +14,7 @@ concurrency:
 
 defaults:
   run:
-    working-directory: tools/src/test/php/1.x
+    working-directory: tools/src/test/php/schema-v1
 
 # https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/controlling-permissions-for-github_token
 permissions: {}
