+++ tools/pom.xml
@@ -195,9 +195,9 @@
 
     <profiles>
         <profile>
-            <id>v1x-tests</id>
+            <id>schema-v1-tests</id>
             <properties>
-                <surefire.includeTags>v1x</surefire.includeTags>
+                <surefire.includeTags>schema-v1</surefire.includeTags>
             </properties>
         </profile>
     </profiles>
