+++ tools/src/test/js/schema-v1/json-schema-lint-tests.js
