+++ tools/src/test/php/schema-v1/json-schema-lint-tests.php
