+++ tools/src/test/proto/schema-v1/buf_breaking-version.yaml
