+++ tools/src/test/proto/schema-v1/test.sh
